LPS-196687 Move mouse dragging hooks to MouseDragAndDropContext
