+++ modules/apps/layout/layout-content-page-editor-web/types/src/main/resources/META-INF/resources/page_editor/plugins/fragments_and_widgets/components/reorder_sets_modal/Item.d.ts
@@ -32,12 +32,4 @@ export declare function ReorderDropdown({
 	numberOfItems,
 	onDropItem,
 }: ItemProps): JSX.Element;
-export declare function useMouseDropTarget(
-	itemId: string,
-	itemIndex: number,
-	onDropItem: ItemProps['onDropItem']
-): {
-	dragOverPosition: DragOverPosition | null | undefined;
-	targetRef: (targetElement: any) => void;
-};
 export {};
